+++ src/main/java/com/nuguna/freview/common/service/impl/PostServiceImpl.java
@@ -1,10 +1,7 @@
 package com.nuguna.freview.common.service.impl;
 
-import com.nuguna.freview.common.dto.response.page.NoticeDetailResponseDTO;
-import com.nuguna.freview.common.mapper.NoticeMapper;
 import com.nuguna.freview.common.mapper.PostMapper;
 import com.nuguna.freview.common.service.PostService;
-import java.sql.Timestamp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,17 +9,10 @@ import org.springframework.stereotype.Service;
 public class PostServiceImpl implements PostService {
 
   private final PostMapper postMapper;
-  private final NoticeMapper noticeMapper;
 
   @Autowired
-  public PostServiceImpl(PostMapper postMapper, NoticeMapper noticeMapper) {
+  public PostServiceImpl(PostMapper postMapper) {
     this.postMapper = postMapper;
-    this.noticeMapper = noticeMapper;
-  }
-
-  @Override
-  public NoticeDetailResponseDTO getNoticeBySeq(Long postSeq) {
-    return noticeMapper.selectNotice(postSeq);
   }
 
   @Override
@@ -37,20 +27,8 @@ public class PostServiceImpl implements PostService {
   }
 
   @Override
-  public boolean updateNotice(Long postSeq, String title, String content, Timestamp now) {
-    int result = noticeMapper.updateNotice(postSeq, title, content, now);
-    return result == 1;
-  }
-
-  @Override
   public boolean deletePost(Long postSeq) {
     int result = postMapper.deletePost(postSeq);
-    return result == 1;
-  }
-
-  @Override
-  public boolean insertNotice(Long postSeq, String title, String content, Timestamp now) {
-    int result = noticeMapper.insertNotice(postSeq, title, content, now);
     return result == 1;
   }
